FEBioStudio: merged GLCOLOR and GLColor classes.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBioStudio@12464 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioStudio/ModelTree.cpp
@@ -538,7 +538,7 @@ void CModelTree::Select(const std::vector<FEObject*>& objList)
 	clearSelection();
 	m_view->SetCurrentItem(-1);
 
-	int N = objList.size();
+	int N = (int)objList.size();
 	if (N == 0) return;
 
 	QTreeWidgetItemIterator it(this);
